django_weather_tracker/weather_tracker/serializers/weather_report_serializer.py
from rest_framework import serializers

from weather_tracker.models import WeatherReport


class WeatherReportSerializer(serializers.ModelSerializer):
    class Meta:
        model = WeatherReport
        fields = "__all__"
