django_weather_tracker/requirements.txt
requests~=2.32.3
python-dateutil~=2.9.0.post0
xmltodict==1.0.2
black==25.1.0
Django==6.0.0
psycopg2-binary==2.9.11
environs[django]==14.5.0
djangorestframework==3.16.1
